requirements.txt
# 🏡 NatPropTech - Requirements
# Autor: MiniMax Agent
# Data: 17 de Novembro de 2025

# Google Cloud AI
google-generativeai>=0.3.0
google-generativeai>=0.3.0
google-ai-generativelanguage>=0.3.0
google-ai-generative-ai>=0.3.0
google-auth>=2.22.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.1

# BigQuery
bigquery-frames>=0.10.0

# Web Framework
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
python-multipart>=0.0.6

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
httpx>=0.24.0

# Development
streamlit>=1.25.0
plotly>=5.15.0